relocalization: the aligner leaves the module
lidar/relocalize.py is now the algorithm - RelocalizeConfig, PreparedMap,
Fix, prepare, align, relocalize - with no streams, no ports, no clock and
no dimos runtime beyond a pydantic base. lidar/module.py keeps what a
module is for: resolving the map file, throttling the cloud stream, and
turning a Fix into a TF.

The eval already only wanted the algorithm, and now says so in its imports.
